Schema JSON file for facades
The following removes the describe-api command. Instead we ensure
that there is a json file in apiserver/facades/schema.json that
represents the current schema of the api server facades.

This file can then be used for pylibjuju to render the client at
any point.

Included in is the ability to verify the schema is up to date and
if they're not, to error out and show the developer what to do.