Add LeaseGroup method to lease stores
This allows getting all leases for a specific namespace/model
combination - it allows us to get the leadership for a model rather
than getting all leases across models and then filtering them
down (which is a lot of extra work when there are hundreds of models).

Implement the interface method for the state, raft and dummy lease
stores.