Intermittent failure CharmProfilingInfo
The following changes ensure that we wait until the
context.KillWithError(err) is called before attempting to check
that worker is killed. This is a logical race, where the mocks
expect that CharmProfilingInfo is called and returns an error, but
it collapses the wait groups. This in turns causes context.Kill(nil)
to be called at the same time. This leads to a race between the
two context kill methods. The tests expect KillWithError(err) to
happen before Kill(nil).

By using the shim we can make the wait group collapse AFTER
KillWithError(err) is called. This removes the logical race between
the two.

It should be noted other tests are probably susceptible to this.