Tie the watcher lifecycle to the catacomb
The machines profile watcher was missing the ability to tie itself
to the provisioner task life-cycle, which means that when a
provisioner task was killed or restarted then the watcher was still
spinning and never cleaned up, leaking as a resource.

There is another problem that @jameinel spotted in the creation of
the provisioner task, is that if the provisioner errored out on
the construction of the provisioner task, then we never killed the
other watchers and they could also be leaked. We should look at
those in the near future to ensure that any watchers that die on
construction are also cleaned up.