Output warning info when series is invalid
To help diagnose issues when the distro-info is out of date, we
print warnings to the command line to help operators understand
what the potential cause of the issue is.

In the PR we print out what warnings where made when parsing the
os series package. This should help operators diagnose the fact their
system is out of date and needs updating.