Bump to the latest charm.v6 dep
The following commit bumps to the latest charm.v6 dep, which takes
the offer url from the juju package and puts it into the charm.v6
dep so that we can easily validate offer urls from the bundle data
perspective without having to duplicate code. With this in mind
we also change the juju code to use that new changes.

Drive by fixes include cleaning up whitespace on lines