Use spec name for zone name
When not in clustered mode and trying to use a different name for
the LXD profile in clouds.yaml, the server name will be wrongly
be reported by lxd as the host name. With recent changes to the
cloud spec and credentials, this will cause LXD to fail bootstrapping.

By over writing the server name with the spec name in non-clustered
environments we should then be allowed to bootstrap at will.