Update to the latest testing lib
When attempting to contact HTTPS url using a HTTP endpoint, in
go version < 1.12 it would return with a malformed garbage error.
In go version >= 1.12 the error is no longer returned and instead
a status code of 400 is returned.

In order to handle this is the correct manor for both past and future
runtimes we should hide this fact and just get prior errors report
that a bad request was sent.