Start tracking Application in cache.
Some updates to start making 'relation-get' work. We no longer high
application and unit data at the Cache level, but instead combine them
down at the command level (so that the user can supply an option
indicating what part of the relation data they want to read.)

Still needs API updates to be able to expose the application data bag,
but all of the code should be roughly structure how we want to consume
it.