Schema gen changes
The following ensures that all properties get flattened to one struct
because that's how go actually manages them when calling json.

This has huge problems for pylibjuju and in turn will cause lots of
issues when it comes to building a new client. I expect to see a lot
of churn.