Better error messages
The following changes how Plugin call back is called. Instead of
terminating inside the Plugin callback, we instead allow more of a
middleware approach, that each callback can be composed. This allows
us to do some very useful things with the return values of the
failed callback.

For example we can use the subcommand name to return a helpful
message like:

```
ERROR juju: "controllerss" is not a juju command. See "juju --help".

Did you mean this?
	controllers
```