Validation of series during bootstrap
The following commit changes the way we bootstrap by validating
against a set of know supported juju series.

Moved the core of the validation into series package in core, so
that we can reuse it from both environs and provider.

Also introduces quick.check to do fuzzy testing of inputs and
outputs.