Get a test FAIL if the package doesn't build.
It seems that when running `go test ./...`
if a subdirectory fails to build, but we have no tests in
that directory the failure we get is:
```
./schemadocs.go:294:10: undefined: strings.ReplaceAll
```

However there isn't anything that clearly indicates a failure. By adding
a dummy Test function we get:
```
./schemadocs.go:294:10: undefined: strings.ReplaceAll
FAIL    github.com/juju/juju/scripts/schemadocs [build failed]
```

Which at least means we can find the issue in the test run output.