Isolated the real cause.
The old check that was waiting to observe that I was no longer the
leader really was the problem. If you put a 1ms delay in the 'stepping
down' code, it could get all the way to WaitAdvance while still having
self as the leader. Which meant that noLeaderCheck would trigger, but it
would see self as still the leader, thus it wouldn't give
ErrNoLeaderTimeout.

The trick to force the failure was:
```github.com/hashicorp/raft/raft.go:543
				} else {
+					time.Sleep(1*time.Millisecond)
					r.logger.Printf("[INFO] raft: Removed ourself, transitioning to follower")
					r.setState(Follower)
				}
```