Refactor AdvancedPackagingConfig to support multiple pkg managers
The original interface definition was strongly coupled to the use of
apt. This is quite evident by the naming conventions that were used but
is a bit misleading as the various cloudinit_xxx implementations would
simply rename the AptProxy and AptMirror settings to something that
would make sense for a particular OS (centos is such an example).

Given that we need to handle setting up proxies for both apt/snap if
the target supports them I have replaced the proxy arguments with an
interface that closely matches the ProxyConfiguration type introduced in
the previous commit. This allows us to delegate the actual
interpretation of the settings to the actual cloudinit implementation
for the target.