Generate proper set of commands to setup snap store proxy via cloudinit
While the proxyupdater worker is already capable of setting up snap to
work with snap store proxies, it runs at a later stage than the point
where we might need to install snap packages (e.g. lxd dependencies).

This commit updates the cloudinit code to include the appropriate
commands so that we can setup snap proxies as early as possible.