Expose GetChangesMapArgs for clients to use
The following exposes GetChangesMapArgs on the Bundle api server,
including adding the methods to the Bundle client. The changes in
the cmd/.../bundle.go requires a LOT of changes, which requires
more time than I've currently got. So instead of doing it in one
stage, I've updated it to make it better than before and left a
very large TODO. Either I'll get time to finish it off at some
point, or someone else will...

In an ideal world, most of the overlay bundle changes would also
be put on the server, again this is a multi-day effort to get it
working and tested, so that again is left as an exercise of the
reader.

What do work at the moment is getting a map of arguments from the
api server, so clients like pylib can work without jumping through
hoops and crashing everytime someone updates bundlechanges.

Note: bundlechanges is a minefield to clients :o