Ensure we set info.Status on update
The following commit ensures that we set the info.status before
performing a store update if the old info is not nil. In other
words, we need to reuse the data from the old info and set it on
the new info which we'll be saving.

The reason for the tests changes, is now that we have the old
data we don't need to save any changes if the data is the same.

This has been tested with pylibjuju and the status changes are
now populated from the start, allowing you to monitor the changes
of a charm.