During Uniter teardown, don't wait for BlockUntilLeadershipReleased
The APIServer isn't tearing down, and the cancel channel doesn't work on
API calls. Instead, we just let the goroutine die on its own.

This only triggers when we have an application with multiple units, and
the Uniter agent for a unit that is not the leader wants to restart.