Force-removing an application with a unit in error
If you had a unit with a hook error, and you tried to remove the
unit's application it would remain. Trying again with force wouldn't
remove it either because the unit was already dying, so the unit
destroy wouldn't be done again. Change the cleanup so that if the
destruction is forced, it propagates the destroy again so that
fallback cleanups can be scheduled.