Protect sends with a check of tomb.Dying()
We were doing bare sends on the out channel if we saw a Changes(), and
if we raced with a tomb.Dying the goroutine could block and never exit.

This was just by inspection, it is unlikely to be visible in the wild,
since you only leak a goroutine if you get a controller change at the
same time you're shutting down.