vsphere: Don't expose acquireMutex outside vsphereclient
Rather than passing it into CreateVirtualMachine (and so requiring the
provider code to know about it, store it on the client struct, set it
to the default acquire function in vsphereclient.Dial and override it
in the tests.

This is much cleaner - I'm blaming jetlag for doing it the other way
first.