Github actions
The following adds github actions for statically checking the code
base per push of a PR.

It uses the new test suite to test static analysis. This does two
things; firstly it helps verify that the tests framework is correctly
setup and secondly that the test framework can fulfill one of its
fundamental goals - run anywhere.