Add the ability to track Resource objects.
Also allow for a default of 30min for the chunks query to timeout (in
case it really is just a slow query). Also allow a flag to disable chunk
processing entirely.

Work out how real "resources" are referenced from Applications and label
them appropriately. We may also need to do something similar for
UnitResources, but for now just look at App references.