Migration runner
The following changes from using global functions to typed migrations
runners. The change is in response to CR feedback where it would be
better to encapsulate types associated with each other on one type
class.

The long term goal of this change is two fold:

 1. Move migration runners out of the state package into their own
    package so that we enforce boundaries on our abstractions.
 2. Allow better coverage of our exporting, without resorting to
    testing using integration for EVERY branch/statement.

There are some sticking points with this change to prevent the package
move, as the inversion of controller around the Attach method uses a
State type that prevents moving it out. Circular dependencies bites us
right at the end, as we can't import the state package if we're already
importing the migration package.