Make import more like export
The following changes attempt to make import more like export.
The problem we have here though is that export can deal with
interfaces and solid types, import can not. Instead we have to
deal with concrete state types. Those state types are db documents
and they have to be created like so.

More thinking about how we model our dependencies when importing
should be done.