Introspect controller on destroy
The following attempts to collate as much information about a controller
when destroying it, so that we can better understand how the controller
behaves when under testing.

To run you need to have specified an artifact otherwise the artifact is
cleaned up. You can do this with the following setup:

     ./main.sh -a output.tar.gz smoke

All files will be namespaced per controller.