CMR Migrations
Included in the commit is the addition of the cmr entities for
consuming a remote application. The code is hidden behind a feature
flag to ensure that we don't break any imports that already exist.

Much more work is required to ensure that we also import all the
other entities required.

Part of the change is also the introduction of go mocks in the state
package. This allows us to reach a higher level of coverage without
the need of the persistence layer.

Not only are the tests in mocks, but additionally added to the existing
integration tests that we have in state.