Systemd uses 'infinity' where ulimit uses 'unlimited'
It is unclear why these use a different syntax, and we could just put
'infinity' in our Service definitions. However, this feels like a
'ulimit' vs 'systemd' thing, so lets encode the Systemd knowledge inside
the Systemd service renderer.

See bug #1845350 for a description of where this comes into play.