Cleanup relation-get arg parsing.
There are several cases we are trying to handle, when the user specifies
an app vs a unit, or they get the defaults without supplying anything.
Now we use the user input first, rather than checking the context first.
No need to check the context if the user is overriding it anyway.
Do a little bit of magic so that 'relation-get' bare with only an
application context sets --app. And if the user specifies '--app' but
passes a Unit name, then we convert that into an app name.