Handle BOOTSTRAP_REUSE_LOCAL
The following allows us to completely reuse a local bootstrap,
so much so, that you can now test the integrations on your local
work. This also doesn't destroy your local controller, only the
models that where created whilst running the tests.

As a driveby I've made sure that the CLI tests pass again.