Ensure we fail at bootstrap
The following changes how we handle verbose logging in side the suite,
instead of pipefail (which isn't available in shell, only bash and
above), we now ensure that `-v` flag turns on debug statements for
bootstrapping.

This was because the CLI suite would fail because underscores where
used for a model name, which isn't allowed. Rather than change juju,
just ensure we fail early.