Rewrite instance poller worker to use only one go-routine
The new instance poller implementation maintains two groups: a short and
a long poll groups.

Entries in the short poll group are polled with (per entry) variable
poll intervals that start from ShortPoll and incrementally get bumped
upwards until the LongPoll interval is reached at which point they will
remain capped to that interval.

On the other hand, entries in the long poll group are always polled at
the LongPoll interval.

The worker uses two timers to decide which group to poll next. It then
iterates the entries in each group and collects the instance IDs for the
entries that are ready to be polled. The instance ID list is passed to
the provider's Instances() method which makes a single call (or emulates
one depending on the provider) to the provider APIs to obtain provider
information about each entry in the batch. We then process each entry
and update the instance status/network information for the machine model
where it's required.

Once an entry of a poll group is processed, the worker checks whether
the entry can be moved to the other group using the following rules:

- if an entry is in the short poll group, has at least one network
address assigned and the machine reports a status of "started": move it
to the long poll group.
- If an entry is in the long poll group and for some reason we were not
able to obtain the instance status from the provider (e.g. the provider
sees the machine as dead but we have not *yet* been notified): move it
to the short poll group.
- If an entry is in the long poll group and we suddenly lost the network
address assignments: move it to the short poll group.

To add new machines to a poll group, the worker watches the model machine
collection for changes. Once a change is received, the worker will first
check if this change affects an already polled machine. If so, the
worker then examines whether the machine is dead and removes it from the
list of polled machines. Otherwise, assuming that something interesting
has happened to the machine, it will be moved (if not already there) to
the short poll group and its poll interfal will be reset so we can poll
its status as soon as possible.

Finally, if the changed machine tag belongs to a machine which the
worker is not yet polling, a new entry will be created for it and placed
in the short poll group.