Update pull request template
To help others remember that we should perform certain tasks, it
might help to give people a gentle nudge in the right direction.

In doing so I've updated the PULL_REQUEST_TEMPLATE.md to include
that info.

Sidenote: we can move this template and the issue template into
the .github root folder in an attempt to keep the root of juju/juju
cleaner.