CAAS unit init container for hook ctx and juju-run
This work moves the prepare logic outside of the command runner
for k8s into an init container that initializes the pod's /var/lib/juju.

The init container shares /var/lib/juju and /usr/bin/juju-run via an
emptyDir volume with the workload containers. When the init container
starts it waits on an abstract domain socket for context to be delivered
via the caasunitinit worker.

The caasunitinit worker watches on the k8s api for the juju-init-pod to
enter the running state. It then uses the exec framework to copy over various
required files such as the charm and config then finally invokes the
command to complete the init.

- Watch on "juju-init-pod" running state.
- Add jujud caas-unit-init command.
- Add caasunitinit worker to caasoperator agent.
- Add "juju-init-pod" container to Workload using operator image.