Prevent instancepoller worker from shutting down due to ErrNoInstances
According to the environs.Networking interface docs, calls to Instances
may return either a ErrPartialInstances if some of the intance IDs were
not found or ErrNoInstances if none of the instance IDs were found. The
latter error is probably rare and will happen if all entries in the poll
group suddenly disappear from the provider. However, we should handle
such a case and prevent the worker from exiting with an error.