Don't trigger relation-changed if there are no app settings.
So if app version == 0, don't trigger a change event.
This is much more compatible with existing charms, and means we don't
have to rewrite the test suite, either. :)
relation-changed app, with no app settings *could* be a replacement for
"relation-joined" but even so, I think we should start conservative.

Still have some Uniter tests failing, but getting closer.