Clean up branches on model destroy
The following cleans up any branches associated with a model, when
the model is destroyed. Ensuring two things:

 1. We remove any orphaned branch models from the db
 2. We ensure that there won't be any conflicts from leaving any
    branches around.

The code attempts to use the latest Apply change strategy, which
applies a visitor pattern to a change operation to clean up the
branches.

It would be nice for these strategies to be moved out from state
and fully mocked, but probably not today!