Update gce provider to support getting net interfaces for multiple instances
The new NetworkInterfaces call signature was wired in a previous commit
to call networkInterfacesForInstance in a loop for each provided
instance ID. The networkInterfacesForInstance would fetch the zone/network
information and then call out to Instances with a single ID as argument.
The returned information would be used to construct the network
interfaces for the instance.

With this commit, Network interfaces will load the shared zone/network
information once, then call Instances with the full set of IDs and
iterate the result list to construct the interface lists for each
instance.

Since this change makes networkInterfacesForInstance redundant, it has
been removed as part of this commit.