Test deploying LXD Profiles
The following deploys a series of lxd profiles in both principal and
subordinate charms.

There is a massive issue around destroying storage on a lxd machine.
Some time should be spent trying to workout why that fails. It's not
like it's a rare occurrence, every other time at least when running

    ./main.sh -v deploy