Add a timeout to hubwatcher sending
If a state-level watcher doesn't unwatch all of the items it has
watched when it stops, all of the other watchers for that model can be
blocked behind it. The flush method will sit there not sending any
changes out, which has the effect that no watchers will fire. This can
be time consuming to track down.

Since the watchers that use the hubwatcher should all be in tight
loops receiving changes, add a 10s timeout to the flush send so that
we can raise an error in the log. Capture the stack for each added
watch and report that in the logged error to indicate the misbehaving
watcher.