Attempt to use github mirror.
The following uses github as a mirror for the flaky golang.org
service. This should attempt to get the source from the same
locations as the other deps and make our dependency woes less
sucky.

More investigation should be done around this as it doesn't look
like the lock file picks up the changes, but essentially this should
work according to the docs.