Patch exported binding descriptions when migrating from 2.6.x controllers
This patching must be performed as 2.6.x controllers export application
bindings as a map where the values are space names whereas 2.7+ exports
maps with space IDs.

Also, 2.6.x controllers use "" as the default space name. We currently
map this to network.DefaultSpaceName but we should update the code to
use the DefaultAlphaSpaceName when the relevant changes land.