Track branch by num of units
This is a naive attempt at adding a tracking branch option to
track the number of units in stages. It's naive in the fact that
it doesn't validate if you just pass a unit type or if you pass
multiple types. So it doesn't accumulate across applications and
units.

That requires some additional feedback to track the number of
items that it has tracked.