Validate -n correctly
To ensure operators correctly enter valid values for the tracking
of a branch, ensure we can verify when they do or don't enter
a valid -n value.

This complication arises because we want to ensure that we don't
accidently allow them to track all branches when they don't want
to.

Also the rise in complexity is shown when we want the inverse of
track all things as a non-value. Most things in juju the default
value is track 1 as the default. Not the case here.