Run all flag
Remove the ability to run all tests suites as default. This prevents
mistakes when attempting to run the test suites and causes the
bootstrap of juju in substraits that can be expensive. To prevent the
feature from being removed completely, we've enabled -A, which allows
the developer to say, they know what they're doing.