Introduce WatchMachineStartTimes watcher
This is a stringwatcher which checks for changes in the (Life,
AgentStartedAt) tuple for a model's machines. This new watcher will be
used by the instance poller in place of the (misleadingly-named)
WatchModelMachines watcher which only reports changes to the machine
Life field.

To ensure that the watcher does not hammer the DB the watcher uses a
batch-based approach. The machine collection is monitored for changes
and the changed document IDs are de-dupped and added to a set. The
watcher maintains a timer which is being reset every time a new change
appears. The timer expiration time is controlled via the quiesceInterval
parameter that is passed to the watcher c-tor. Once the timer expires,
the collected document IDs are passed to a mongo query which selects the
(Lide, AgentStartedAt) fields, compares them to the last-seen value and
retains the ones that have changed. Additionally, dead machines are
automatically removed from the last-seen value map to reclaim memory.

The retained set of document IDs are made available to the watch channel
consumer. Note that if additional changes appear before the consumer
manages to read off the watcher channel, they will be appended to the
existing change set and returned as a single batch.