Wire instancepoller to invoke WatchModelMachineStartTimes
This change allows the instancepoller to be notified when the machine
agent restarts so it can eagerly inspect the provider instance
information to discover changes to shadow IPs.

This can happen for instance on EC2 if a machine instance is stopped and
then re-started.

A side-effect of the current way of doing things is that if the machine
was placed in the long polling group, we wouldn't detect shadow IP
changes until the machine gets polled again (~10-15min). This detection
delay prevens us from SSHing into rebooted machines and also seems to
block `juju upgrade-controller` until the new IP is detected.