Process relation phase
The following adds a new model migration phase. This phase allows us
to identify a placeholder in the migration statemachine for processing
any relations needed.

This should then allow us to correctly update remote applications,
offer ids and firewalls in the future.

The code changes the phase migration steps in core and changes where
states can correctly move to.

This changes the migration master, but minions shouldn't be affected
as they don't error out on new phases (default is a no-op).

I've tested this locally and it works as designed.