Collect and emit initial machine set in machine start time watcher
Some of the tests in the jujud/agent package are written with the
assumption that the machine watcher (see
apiserver/commmon/modelmachinewatcher.go) will *always* emit an initial
change with all known machine models. These tests would wait for the
first set of changes to be emitted *before* mutating the machine state.
This causes issue with the batching approach used by the new machine
model start time watcher.

This commit addresses the issue by ensuring that the watcher always
emits the initial (even if it's empty) set of machines.