Update InterfaceInfo to support multiple addresses/shadow addresses
Given that we may potentially have multiple addresses for an interface,
the general convention is to place the *primary* address at the top of
the address list. The PrimaryAddress method on the InterfaceInfo can be
used to fetch the primary address (if defined).

The majority of providers only support a single address so this change
is backwards-compatible. The ec2 and gce providers in particular seem to
support assigning multiple private IPs to a NIC although only one can be
flagged as the primary.