Add helpers for infering space from address or CIDR+provider subnetID
The infer logic for addresses works by parsing the subnet CIDRs and
attempting to match the specified address to a single CIDR (multiple
matches cause an error to be returned).

The expectation is that the code using the infer helpers will obtain the
SpaceInfos once and then iterate on a list of addresses. In order to
avoid having to parse CIDRs every single time, this commit tweaks
network.SubnetInfo so that it memoizes parsed CIDRs and adds the
ParsedCIDRNetwork helper to get the parsed CIDR instance.