Add instance poller facade method for setting network information
The implementation of the new API method attempts to infer the space
information for the set of addresses provided by the instance poller and
is also capable of handling multiple Address/ShadowAddress entries.

If the space-inference attempt fails with ErrNotFound (in other words:
no CIDR-related conflict was detected), the code will automatically
assign the alpha space as a fall-back. The idea here is that the juju
CLI will provide the appropriate commands for moving IPs to different
spaces.

For shadow addresses, we try to infer the space name from the address
value only. If that fails (e.g. publicly routable address), we
auto-assign the alpha space.