Update the instancepoller to send full interface info to the controller
If the underlying provider supports environ.Networking (the
NetworkInterfaces call specifically), the instance poller will query it
with the same instance ID batch as the one passed to Instances(). On the
other hand if the provider does not support this method, we query the
list of provider addresses from the Instances() output and coerce it
into a slice of network.InterfaceInfo.

When following the latter approach, we obviously lose some fidelity as
we cannot detect whether a public IP is a real IP attached to an
interface or a shadow IP. Instead, a separate interface will be
generated for each private/shadow address entry.

The list of network interfaces (for each instance) is sent to the
controller via the new SetProviderNetworkConfig which attempts to infer
the space ID for the interface private/shadow addresses, updates the
machine provider addresses (if modified) and returns back the current
set of provider addresses and a flag to indicate whether the machine
addresses were modified.