Unit test import application offer
The following code unit tests the importing of application offers. It
unit tests at the strategy level by using mocks. It should give us a
good bit of coverage of the imporint side of things.

There are a few gaps where the description package isn't giving us
everything we need.