Add JUJU_HOOK_NAME to hook environments
There are currently indirect mechanisms for a charm to identify the hook
name (argv[0] or JUJU_CONTEXT_ID which does not have a publicly
documented format).

This change adds it to the hook context and environment variables.

JUJU_HOOK_NAME will be set to a function name for functions.

See LP: #1854505