Get juju output without calling model command
The following prevents an issue where the pylib code requests a model
before asking to switch. This can cause issues where it expects the
model to be there, even though we just unregistered it.

So the code now calls into the "global" juju commands, as we don't have
any other controllers (it checks it prior to calling this code) and so
it's safe to call `juju switch` and see the output correctly.

The code is actually doing what we expect here, rather than calling
`juju models` first.