Rework description package exporting
The following re-works the description package exporting methods. This
just makes it very easy to marshall between the state package and the
description package, without having the need to jump between types for
no point.

Validation is done when we make the state types, so the additional
marshalling isn't actually a must, so it's a general clean up with
additional tests.