Import endpoint maps
The following attempts to import the endpoints map, which is new, that
allows us to use both the internal and public facing end point name.
This should now allow us to correctly wire up the offers and offer
connections to the right applications.

It's worth noting that this doesn't pass atm because the tests aren't
correctly wired up. That will come, once I figure out how to do that.

Additionally I've updated all the mocks, which is why they have churned,
but they're just a mechanical change that adds a function call to the
test helper.