Import application offer endpoints
The following code brings in the new offer endpoints description package
changes. The changes now manifest themselves as a map[string]string,
rather than a []string type. With these changes we can get hold of the
internal and external namings of a endpoint.

This should now allow use to correctly import a application offer. More
is yet to come though, as we need to correctly import the ACL. Once
that's done I can call this part DONE!