Revise random folder generation
The follow changes how we generate our random folder names. Instead of
piping all of /dev/urandom to the pipe, use head to get 10 lines and
then work on that.

For some reason the original command would just hang on the head
section:

>  `cat /dev/urandom | tr -dc 'a-z0-9' | fold -w 8 | head -n 1`

I have no idea why this hangs, I can't get it to hang on any other
distro release, so instead of spending hours on this, I'm going to chalk
it up as a quirk and do it another way.

That new way asks for the head (first ten lines) of urandom then works
on it.

The new way feels much more efficient, but I've done no tests to prove
this!