Fix intermittent failure in ActionSuite.TestFindActionTagsByLegacyId
It repeatedly enqueues actions until it gets a uuid that starts with a
digit, then switches to integer ids and tries to ensure a collision
with the first digit of the uuid (to make sure the querying avoids
them). But if the first digit of the uuid was a 0 it would fail
because the numeric ids start from 1 so no actions were generated.