External controllers
The following exports and imports external controllers. There has been
some effort in trying to ensure that we import the external controllers
correctly, by enforcing vertical boundaries around state and description
packages.

It should now be possible to correctly import and export a external
controller. There requires some work to ensure that we've not broken
backwards compatibility by attempting to import a migration yaml,
without a remoteApplication.

In theory if a remoteApplication is present we prevent the import
happening. In reality we probably need to check for the existence of
other types, but that ship has already saled. The very nature of all the
software that we need to add these checks to, are already released!