Mechanical changes for Firewall Rules
The firewall rules importing task was wrong, most if not all the
importing logic, doesn't validate what's comming in and assumes because
the newly created model is a blank slate, that it can just insert
documents without retry logic.

When making this change to be in line with all the other tasks, I put
the firewallrules type methods back to what they where originally. It
doesn't and shouldn't need to know about migration and makes for a
sloppy firewall service.

In doing these changes I added missing tests for the migration import
tasks, so we can at least verify that our import script works for both
integration and unit level testing.