Call "dispatch" hook first, if doesn't exist use hook-name.
If the "dispatch" file is not found, try to find hook-name instead.
Leave the rest of the runner the same.  "dispatch" is a special
hook to replace individual hooks.  Juju charms are in a state of change,
so we to cover the old and new methods for now.

Add JUJU_DISPATCH_PATH EnvVar to be [actions|hooks]/hook-name.