Speeds up the loading of the units by caching the opened ports.
The opened ports are kept on documents associated with the machines and the subnet.
All the current code around dealing with opened ports in the allwatcher assume
an empty subnet ID. This will need to be updated as we start work on the limited
exposing of ports to subnets.

As the model is processed it sets the model type in the context.