Trusted bundles
Move trusted bundles from acceptance tests to integration tests. As part
of the CI work, we're moving from acceptance tests to integration tests,
this should hopefully focus our efforts in one location, rather than
multiple.

As this going to be a long process, I'm tackling the bitesized tests
first. This just happens to be small enough in size!