Introduce the multiwatcher factory to the apiserver.
This work defines a new interface that lives in the core/multiwatcher package
that is used to create multiwatchers for either a model or a controller.

This multiwatcher factory is threaded through the facade context to make it
available to call, and is used by the client and controller facades to create
the multiwatchers.

A shim is temporarily created in the apiserver package to provide this
interface with the current state methods. This will be removed in a followup
branch that brings in the multiwatcher worker itself.