Update the state package to match worker/multiwatcher expectations.
As the multiwatcher work continues, we are moving towards just having
one type of all watcher backing. This is then used by the multiwatcher
worker to provide the difference filters.

The all watcher that watches all the models has been renamed to better
reflect its role, and collections that were added to the one model watcher
have been added to the other.